+++ src/OpenTelemetry.Instrumentation.Hangfire/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Added client side instrumentation for jobs ([#421](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/421))
+
 ## 1.0.0-beta.1
 
 * Updated OTel SDK package version to 1.2.0